# Remove all unsafe from codebase, improve parsing speed