--- a/fuzz/fuzz_targets/fuzz_target_1.rs
+++ b/fuzz/fuzz_targets/fuzz_target_1.rs
@@ -11,6 +11,7 @@ fuzz_target!(|data: &[u8]| {
         if let Ok(d) = parse_http_date(s) {
             let o = fmt_http_date(d);
             assert!(!o.is_empty());
+            assert_eq!(parse_http_date(&o).expect("formatting to round trip"), d);
         }
     }
 });