--- a/src/date.rs
+++ b/src/date.rs
@@ -1,6 +1,5 @@
 use std::cmp;
 use std::fmt::{self, Display, Formatter};
-use std::num::ParseIntError;
 use std::str::{from_utf8_unchecked, FromStr};
 use std::time::{Duration, SystemTime, UNIX_EPOCH};
 
@@ -244,27 +243,40 @@ impl PartialOrd for HttpDate {
     }
 }
 
-/// Convert &[u8] to &str with zero checks.
-///
-/// For internal use only.
-/// Intended to be used with ASCII-only strings.
-fn conv(s: &[u8]) -> &str {
-    unsafe { from_utf8_unchecked(s) }
+fn toint_2(s: &[u8]) -> Result<u8, Error> {
+    let high = s[0].wrapping_sub(b'0');
+    let low = s[1].wrapping_sub(b'0');
+
+    if high < 10 && low < 10 {
+        Ok(high * 10 + low)
+    } else {
+        Err(Error(()))
+    }
 }
 
-fn cvt_err(_: ParseIntError) -> Error {
-    Error(())
+fn toint_4(s: &[u8]) -> Result<u16, Error> {
+    let a = u16::from(s[0].wrapping_sub(b'0'));
+    let b = u16::from(s[1].wrapping_sub(b'0'));
+    let c = u16::from(s[2].wrapping_sub(b'0'));
+    let d = u16::from(s[3].wrapping_sub(b'0'));
+    
+    if a < 10 && b < 10 && c < 10 && d < 10 {
+        Ok(a * 1000 + b * 100 + c * 10 + d)
+    } else {
+        Err(Error(()))
+    }
 }
+
 fn parse_imf_fixdate(s: &[u8]) -> Result<HttpDate, Error> {
     // Example: `Sun, 06 Nov 1994 08:49:37 GMT`
     if s.len() != 29 || &s[25..] != b" GMT" || s[16] != b' ' || s[19] != b':' || s[22] != b':' {
         return Err(Error(()));
     }
     Ok(HttpDate {
-        sec: conv(&s[23..25]).parse().map_err(cvt_err)?,
-        min: conv(&s[20..22]).parse().map_err(cvt_err)?,
-        hour: conv(&s[17..19]).parse().map_err(cvt_err)?,
-        day: conv(&s[5..7]).parse().map_err(cvt_err)?,
+        sec: toint_2(&s[23..25])?,
+        min: toint_2(&s[20..22])?,
+        hour: toint_2(&s[17..19])?,
+        day: toint_2(&s[5..7])?,
         mon: match &s[7..12] {
             b" Jan " => 1,
             b" Feb " => 2,
@@ -280,7 +292,7 @@ fn parse_imf_fixdate(s: &[u8]) -> Result<HttpDate, Error> {
             b" Dec " => 12,
             _ => return Err(Error(())),
         },
-        year: conv(&s[12..16]).parse().map_err(cvt_err)?,
+        year: toint_4(&s[12..16])?,
         wday: match &s[..5] {
             b"Mon, " => 1,
             b"Tue, " => 2,
@@ -317,17 +329,17 @@ fn parse_rfc850_date(s: &[u8]) -> Result<HttpDate, Error> {
     if s.len() != 22 || s[12] != b':' || s[15] != b':' || &s[18..22] != b" GMT" {
         return Err(Error(()));
     }
-    let mut year = conv(&s[7..9]).parse::<u16>().map_err(cvt_err)?;
+    let mut year = u16::from(toint_2(&s[7..9])?);
     if year < 70 {
         year += 2000;
     } else {
         year += 1900;
     }
     Ok(HttpDate {
-        sec: conv(&s[16..18]).parse().map_err(cvt_err)?,
-        min: conv(&s[13..15]).parse().map_err(cvt_err)?,
-        hour: conv(&s[10..12]).parse().map_err(cvt_err)?,
-        day: conv(&s[0..2]).parse().map_err(cvt_err)?,
+        sec: toint_2(&s[16..18])?,
+        min: toint_2(&s[13..15])?,
+        hour: toint_2(&s[10..12])?,
+        day: toint_2(&s[0..2])?,
         mon: match &s[2..7] {
             b"-Jan-" => 1,
             b"-Feb-" => 2,
@@ -348,20 +360,27 @@ fn parse_rfc850_date(s: &[u8]) -> Result<HttpDate, Error> {
     })
 }
 
+fn toint_1(x: u8) -> Result<u8, Error> {
+    let result = x.wrapping_sub(b'0');
+    if result < 10 {
+        Ok(result)
+    } else {
+        Err(Error(()))
+    }
+}
+
 fn parse_asctime(s: &[u8]) -> Result<HttpDate, Error> {
     // Example: `Sun Nov  6 08:49:37 1994`
     if s.len() != 24 || s[10] != b' ' || s[13] != b':' || s[16] != b':' || s[19] != b' ' {
         return Err(Error(()));
     }
     Ok(HttpDate {
-        sec: conv(&s[17..19]).parse().map_err(cvt_err)?,
-        min: conv(&s[14..16]).parse().map_err(cvt_err)?,
-        hour: conv(&s[11..13]).parse().map_err(cvt_err)?,
+        sec: toint_2(&s[17..19])?,
+        min: toint_2(&s[14..16])?,
+        hour: toint_2(&s[11..13])?,
         day: {
             let x = &s[8..10];
-            conv(if x[0] == b' ' { &x[1..2] } else { x })
-                .parse()
-                .map_err(cvt_err)?
+            {if x[0] == b' ' { toint_1(x[1]) } else { toint_2(x) }}?
         },
         mon: match &s[4..8] {
             b"Jan " => 1,
@@ -378,7 +397,7 @@ fn parse_asctime(s: &[u8]) -> Result<HttpDate, Error> {
             b"Dec " => 12,
             _ => return Err(Error(())),
         },
-        year: conv(&s[20..24]).parse().map_err(cvt_err)?,
+        year: toint_4(&s[20..24])?,
         wday: match &s[0..4] {
             b"Mon " => 1,
             b"Tue " => 2,