--- a/src/lib.rs
+++ b/src/lib.rs
@@ -16,6 +16,7 @@
 //! and format timestamps. Convert a sytem time to `HttpDate` and vice versa.
 //! The `HttpType` (8 bytes) is smaller than `SystemTime` (16 bytes) and
 //! using the display impl avoids a temporary allocation.
+#![forbid(unsafe_code)]
 
 use std::error;
 use std::fmt::{self, Display, Formatter};